Do not show code for tree decision plot.
Use %run instead.
